【WA】[AtCoder Beginner Contest 137]C - Green Bin

+++ abc137/abc137_c.cpp
@@ -7,7 +7,7 @@ typedef long long ll;
 
 int main() {
 
-    int n;
+    ll n;
     cin >> n;
     map<string, int> strs;
